refactor: 폴더 구조 재구성
- /pages/auth > LoginPage, SignupPage 포함
