+++ src/components/layout/MainLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import Header from '../common/Header';
-import Footer from '../common/Footer';
+import Header from './Header';
+import Footer from './Footer';
 
 const MainLayout = () => {
   return (
