+++ src/pages/auth/components/AuthPrompt.tsx
