+++ src/pages/auth/components/OAuthSection.tsx
@@ -1,5 +1,5 @@
-import googleLogin from '../../assets/images/google-login.svg';
-import kakaoLogin from '../../assets/images/kakao-login.svg';
+import googleLogin from '../../../assets/images/google-login.svg';
+import kakaoLogin from '../../../assets/images/kakao-login.svg';
 
 const OAuthSection = () => {
   return (
